test: add LRU test
