+++ packages/core/src/shared/index.ts
@@ -14,32 +14,32 @@
  * limitations under the License.
  */
 
-export { checkIfMove, MOVE_BUFFER_VALUE, ROTATE_BUFFER_VALUE } from './check-if-move';
+export { afterInitApply } from './after-init-apply';
 export * from './array-search';
 export * from './blob';
+export { checkIfMove, MOVE_BUFFER_VALUE, ROTATE_BUFFER_VALUE } from './check-if-move';
+export * from './clipboard';
 export * from './color/color';
-export { ColorKit, COLORS, RGB_PAREN, RGBA_PAREN, type IRgbColor } from './color/color-kit';
+export { ColorKit, COLORS, type IRgbColor, RGB_PAREN, RGBA_PAREN } from './color/color-kit';
 export * from './command-enum';
 export * from './common';
 export * from './compare';
+export * from './debounce';
 export * from './doc-tool';
 export * from './generate';
 export * from './hash-algorithm';
 export * from './lifecycle';
 export * from './locale';
-export * from './lru/index';
+export { LRUHelper, LRUMap } from './lru/lru-map';
+export { numfmt } from './numfmt';
 export * from './object-matrix';
+export { queryObjectMatrix } from './object-matrix-query';
+export { moveRangeByOffset } from './range';
 export * from './rectangle';
 export { RefAlias } from './ref-alias';
 export * from './row-col-iter';
 export * from './sequence';
-export * from './sort-rules';
 export * from './shape';
+export * from './sort-rules';
 export * from './tools';
 export * from './types';
-export * from './debounce';
-export * from './clipboard';
-export { numfmt } from './numfmt';
-export { queryObjectMatrix } from './object-matrix-query';
-export { moveRangeByOffset } from './range';
-export { afterInitApply } from './after-init-apply';
